Allow bibtex to be added from arxiv id.
This allows the user to add a reference via an arXiv ID similarly to how a reference can be added
from a DOI or ISBN.  If the arXiv ID has a DOI associated with it (according to the arXiv server),
the DOI will be used.  If it does not (perhaps the paper is unpublished), then a bibtex entry will
automatically be generated from the reference's metadata.

Note that a potential issue with this addition is that if a paper is added before it is published
(i.e., there is no DOI associated with it), and the paper is later published, the updated
information will have to be manually added.